+++ src/components/common/provider-idle-screen/provider-idle-screen.tsx
@@ -1,31 +1,36 @@
 import type { EventEmitter } from '@stencil/core';
 import { Component, Event, Fragment, h, Prop } from '@stencil/core';
+import { getProviderButtonIcon } from 'components/functional/unlock-panel/helpers';
 import { SidePanelHeaderSlotEnum } from 'components/visual/side-panel/components/side-panel-header/side-panel-header.types';
 import type { IProviderBase } from 'types/provider.types';
 import { ProviderTypeEnum } from 'types/provider.types';
 
-import { getProviderButtonIcon } from '../../helpers';
-
 const getProviderIntroText = (providerType?: IProviderBase['type']) => {
   switch (providerType) {
     case ProviderTypeEnum.extension:
-      return 'Open the MultiversX Browser Extension to connect to your wallet.';
+      return 'Open the MultiversX Browser Extension to sign the transaction.';
     case ProviderTypeEnum.metamask:
-      return 'Open the MetaMask Browser Extension to connect to your wallet.';
+      return 'Open the Metamask Browser Extension to sign the transaction.';
     case ProviderTypeEnum.passkey:
-      return 'Use your predefined passkey to connect to your wallet.';
+      return 'Use your predefined passkey to sign the transaction.';
+    case ProviderTypeEnum.walletConnect:
+      return 'Open Xportal to sign the transaction.';
     case ProviderTypeEnum.crossWindow:
-      return 'Follow the steps on MultiversX Web Wallet to connect to your wallet.';
+      return 'Go to MultiversX Web Wallet to sign the transaction.';
+    default:
+      return 'Go to your connected provider to sign the transaction.';
   }
 };
 
 @Component({
-  tag: 'mvx-unlock-provider-intro',
-  styleUrl: 'unlock-provider-intro.scss',
+  tag: 'mvx-provider-idle-screen',
+  styleUrl: 'provider-idle-screen.scss',
   shadow: true,
 })
-export class UnlockProviderIntro {
+export class ProviderIdleScreen {
   @Prop() provider: IProviderBase | null = null;
+  @Prop() introText: string = '';
+
   @Event({ composed: false, bubbles: false }) close: EventEmitter;
   @Event({ composed: false, bubbles: false }) access: EventEmitter;
 
@@ -37,7 +42,7 @@ export class UnlockProviderIntro {
     const extensionProviderIconHeight = extensionProviderIconBaseSize + (10 / 100) * extensionProviderIconBaseSize;
 
     const providerIntroIcon = getProviderButtonIcon(providerType);
-    const providerIntroText = getProviderIntroText(providerType);
+    const providerIntroText = this.introText || getProviderIntroText(providerType);
 
     if (!this.provider) {
       return null;
@@ -70,8 +75,9 @@ export class UnlockProviderIntro {
             <div class="unlock-provider-intro-icon">{providerIntroIcon}</div>
           )}
 
-          <div class="unlock-provider-intro-title">Requesting Connection</div>
+          <div class="unlock-provider-intro-title">{this.introText ? 'Sign transaction' : 'Requesting Connection'}</div>
           {providerIntroText && <div class="unlock-provider-intro-text">{providerIntroText}</div>}
+          <slot name="close-button" />
         </div>
       </Fragment>
     );